Generate resource_population with custom code split
