+++ internal/service/sso/resource_population_gen.go
@@ -7,18 +7,14 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"regexp"
 
-	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/patrickcping/pingone-go-sdk-v2/management"
-	"github.com/patrickcping/pingone-go-sdk-v2/pingone/model"
-	client "github.com/pingidentity/terraform-provider-pingone/internal/client"
+	"github.com/pingidentity/terraform-provider-pingone/internal/client"
 	"github.com/pingidentity/terraform-provider-pingone/internal/framework"
 	"github.com/pingidentity/terraform-provider-pingone/internal/framework/customtypes/pingonetypes"
 	"github.com/pingidentity/terraform-provider-pingone/internal/sdk"
@@ -79,6 +75,7 @@ type populationResourceModel struct {
 	Id               pingonetypes.ResourceIDValue `tfsdk:"id"`
 	Name             types.String                 `tfsdk:"name"`
 	PasswordPolicyId pingonetypes.ResourceIDValue `tfsdk:"password_policy_id"`
+	UserCount        types.Int32                  `tfsdk:"user_count"`
 }
 
 func (r *populationResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
@@ -96,14 +93,15 @@ func (r *populationResource) Schema(ctx context.Context, req resource.SchemaRequ
 			"name": schema.StringAttribute{
 				Required:    true,
 				Description: "A string that specifies the population name, which must be provided and must be unique within an environment.",
-				Validators: []validator.String{
-					stringvalidator.LengthAtLeast(1),
-				},
 			},
 			"password_policy_id": schema.StringAttribute{
 				Description: framework.SchemaAttributeDescriptionFromMarkdown("A string that specifies the ID of a password policy to assign to the population.  Must be a valid PingOne resource ID.").Description,
 				Optional:    true,
 				CustomType:  pingonetypes.ResourceIDType{},
+			},
+			"user_count": schema.Int32Attribute{
+				Computed:    true,
+				Description: "The number of users that belong to the population",
 			},
 		},
 	}
@@ -115,18 +113,18 @@ func (model *populationResourceModel) buildClientStruct() (*management.Populatio
 	result.Description = model.Description.ValueStringPointer()
 	// name
 	result.Name = model.Name.ValueString()
-	// password_policy
+	// password_policy_id
 	if !model.PasswordPolicyId.IsNull() {
-		passwordPolicyValue := &management.PopulationPasswordPolicy{
+		result.PasswordPolicy = &management.PopulationPasswordPolicy{
 			Id: model.PasswordPolicyId.ValueString(),
 		}
-		result.PasswordPolicy = passwordPolicyValue
 	}
 
 	return result, nil
 }
 
 func (state *populationResourceModel) readClientResponse(response *management.Population) diag.Diagnostics {
+	var respDiags diag.Diagnostics
 	// description
 	state.Description = types.StringPointerValue(response.Description)
 	// id
@@ -142,7 +140,9 @@ func (state *populationResourceModel) readClientResponse(response *management.Po
 		passwordPolicyIdValue = framework.PingOneResourceIDToTF(response.PasswordPolicy.Id)
 	}
 	state.PasswordPolicyId = passwordPolicyIdValue
-	return nil
+	// user_count
+	state.UserCount = types.Int32PointerValue(response.UserCount)
+	return respDiags
 }
 
 func (r *populationResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
@@ -382,198 +382,4 @@ func (r *populationResource) ImportState(ctx context.Context, req resource.Impor
 
 		resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root(pathKey), attributes[idComponent.Label])...)
 	}
-}
-
-func populationDeleteCustomErrorHandler(r *http.Response, p1Error *model.P1Error) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	if p1Error != nil {
-		// Env must contain at least one population
-		if details, ok := p1Error.GetDetailsOk(); ok && details != nil && len(details) > 0 {
-			if code, ok := details[0].GetCodeOk(); ok && *code == "CONSTRAINT_VIOLATION" {
-				if message, ok := details[0].GetMessageOk(); ok {
-					m, err := regexp.MatchString(`must contain at least one population`, *message)
-					if err == nil && m {
-						diags.AddWarning(
-							"Constraint violation",
-							fmt.Sprintf("A constraint violation error was encountered: %s\n\nThe population has been removed from Terraform state, but has been left in place in the environment.", p1Error.GetMessage()),
-						)
-
-						return diags
-					}
-				}
-			}
-		}
-	}
-
-	return diags
-}
-
-func (r *populationResource) hasUsersAssigned(ctx context.Context, environmentID, populationID string) (bool, diag.Diagnostics) {
-	var diags diag.Diagnostics
-
-	users, d := r.readUsers(ctx, environmentID, populationID)
-	diags.Append(d...)
-	if diags.HasError() {
-		return false, diags
-	}
-
-	if len(users) > 0 {
-		return true, diags
-	}
-
-	return false, diags
-}
-
-func (r *populationResource) readUsers(ctx context.Context, environmentID, populationID string) ([]management.User, diag.Diagnostics) {
-	var diags diag.Diagnostics
-
-	m, err := regexp.MatchString(verify.P1ResourceIDRegexpFullString.String(), populationID)
-	if err != nil {
-		diags.AddError(
-			"Population ID validation",
-			fmt.Sprintf("An error occurred while validating the population ID: %s", err.Error()),
-		)
-		return nil, diags
-	}
-
-	if m {
-
-		scimFilter := fmt.Sprintf(`population.id eq "%s"`, populationID)
-
-		// Run the API call
-		var users []management.User
-		diags.Append(framework.ParseResponse(
-			ctx,
-
-			func() (any, *http.Response, error) {
-				pagedIterator := r.Client.ManagementAPIClient.UsersApi.ReadAllUsers(ctx, environmentID).Filter(scimFilter).Execute()
-
-				var initialHttpResponse *http.Response
-
-				foundUsers := make([]management.User, 0)
-
-				for pageCursor, err := range pagedIterator {
-					if err != nil {
-						return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, nil, pageCursor.HTTPResponse, err)
-					}
-
-					if initialHttpResponse == nil {
-						initialHttpResponse = pageCursor.HTTPResponse
-					}
-
-					if pageCursor.EntityArray.Embedded != nil && pageCursor.EntityArray.Embedded.Users != nil {
-						foundUsers = append(foundUsers, pageCursor.EntityArray.Embedded.GetUsers()...)
-					}
-				}
-
-				return foundUsers, initialHttpResponse, nil
-			},
-			"ReadAllUsers",
-			framework.DefaultCustomError,
-			sdk.DefaultCreateReadRetryable,
-			&users,
-		)...)
-
-		if diags.HasError() {
-			return nil, diags
-		}
-
-		return users, nil
-	}
-
-	if r.options.Population.ContainsUsersForceDelete {
-		diags.AddError(
-			"Data protection notice",
-			fmt.Sprintf("For data protection reasons, it could not be determined whether users exist in the population %[2]s in environment %[1]s. Any users in this population will not be deleted.", environmentID, populationID),
-		)
-	}
-
-	return nil, diags
-}
-
-func (r *populationResource) checkEnvironmentControls(ctx context.Context, environmentID, populationID string) (bool, diag.Diagnostics) {
-	var diags diag.Diagnostics
-
-	if r.options.Population.ContainsUsersForceDelete {
-		// Check if the environment is a sandbox type.  We'll only delete users in sandbox environments
-		var environmentResponse *management.Environment
-		diags.Append(framework.ParseResponse(
-			ctx,
-
-			func() (any, *http.Response, error) {
-				fO, fR, fErr := r.Client.ManagementAPIClient.EnvironmentsApi.ReadOneEnvironment(ctx, environmentID).Execute()
-				return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, fO, fR, fErr)
-			},
-			"ReadOneEnvironment-DeletePopulation",
-			framework.DefaultCustomError,
-			nil,
-			&environmentResponse,
-		)...)
-		if diags.HasError() {
-			return false, diags
-		}
-
-		if v, ok := environmentResponse.GetTypeOk(); ok && *v == management.ENUMENVIRONMENTTYPE_SANDBOX {
-			return true, diags
-		} else {
-			diags.AddWarning(
-				"Data protection notice",
-				fmt.Sprintf("For data protection reasons, the provider configuration `global_options.population.contains_users_force_delete` has no effect on environment ID %[1]s as it has a type set to `PRODUCTION`.  Users in this population will not be deleted.\n"+
-					"If you wish to force delete population %[2]s in environment %[1]s, please review and remove user data manually.", environmentID, populationID),
-			)
-		}
-	}
-
-	return false, diags
-}
-
-func (r *populationResource) checkControlsAndDeletePopulationUsers(ctx context.Context, environmentID, populationID string) diag.Diagnostics {
-	var diags diag.Diagnostics
-
-	environmentControlsOk, d := r.checkEnvironmentControls(ctx, environmentID, populationID)
-	diags.Append(d...)
-	if diags.HasError() {
-		return diags
-	}
-
-	if environmentControlsOk {
-
-		loopCounter := 1
-		for loopCounter > 0 {
-
-			users, d := r.readUsers(ctx, environmentID, populationID)
-			diags.Append(d...)
-			if diags.HasError() {
-				return diags
-			}
-
-			// DELETE USERS
-			if len(users) == 0 {
-				break
-			} else {
-				for _, user := range users {
-					var entityArray *management.EntityArray
-					diags.Append(framework.ParseResponse(
-						ctx,
-
-						func() (any, *http.Response, error) {
-							fR, fErr := r.Client.ManagementAPIClient.UsersApi.DeleteUser(ctx, environmentID, user.GetId()).Execute()
-							return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, nil, fR, fErr)
-						},
-						"DeleteUser-DeletePopulation",
-						framework.DefaultCustomError,
-						nil,
-						&entityArray,
-					)...)
-
-					if diags.HasError() {
-						return diags
-					}
-				}
-			}
-		}
-	}
-
-	return diags
 }
